RestAPI: Move to Skills.OpenApi package (#887)
### Description
Moving the RestAPI* classes back into the Skills.OpenAPI package, where
they can remain as internals. This aligns with the Skills.Grpc format as
well.
