+++ dotnet/src/Skills/Skills.OpenAPI/Model/RestApiOperationParameterLocation.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
-namespace Microsoft.SemanticKernel.Connectors.WebApi.Rest.Model;
+namespace Microsoft.SemanticKernel.Skills.OpenAPI.Model;
 
 /// <summary>
 /// The REST API operation parameter location.
